+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 4.2.0 (2024-08-06)
+- Feat: Added card holder information in create token and create authentication
+
+## 4.1.2 (2024-05-27)
+- Fix JCB card detection and expiry date
+
 ## 4.1.1 (20240-01-26)
 - Support mid_label in create token and unbundled create authentication